Update README.md
You can see from the Issues page and Google that I am not the only one that thought you had to compile the project in order to eventually get it to run, which is a major headache. I think the proposed change makes it much more clear that you only need to download the executable.